(#36) fix: useSchoolsApi 하나로 합침
